Move name resolvers next to their drivers
name_resolver.go had grown to 1854 lines holding 40 resolver types for
40 different providers, making it a merge-conflict hotspot: any change
touching a provider's endpoints had to edit both its driver file and
this shared one.

Relocate each <provider>NameResolver, its constructor and its helpers
into pkg/accessreview/drivers/<provider>.go, following the convention
grafana, metabase, signoz, posthog and tailscale already used. The file
retains only the NameResolver interface, ErrTerminalNameResolution and
nameStatusError.

This is a pure move: all 82 distinct URL literals are byte-identical
and no cassette changed.
