+++ pkg/accessreview/drivers/better_stack.go
@@ -233,3 +233,18 @@ func parseBetterStackTimestamp(value string) (time.Time, bool) {
 
 	return t, true
 }
+
+// betterStackNameResolver returns the Better Stack team name captured when
+// the API-key connector was created. The team name is the human-readable
+// instance identifier, so no HTTP call is required.
+type betterStackNameResolver struct {
+	teamName string
+}
+
+func NewBetterStackNameResolver(teamName string) NameResolver {
+	return &betterStackNameResolver{teamName: teamName}
+}
+
+func (r *betterStackNameResolver) ResolveInstanceName(_ context.Context) (string, error) {
+	return r.teamName, nil
+}
