+++ pkg/accessreview/drivers/google_analytics.go
@@ -354,3 +354,55 @@ func googleAnalyticsRecords(members map[string]*googleAnalyticsMember) []Account
 
 	return records
 }
+
+// googleAnalyticsNameResolver resolves a GA4 account's display name.
+type googleAnalyticsNameResolver struct {
+	httpClient *http.Client
+	accountID  string
+}
+
+func NewGoogleAnalyticsNameResolver(httpClient *http.Client, accountID string) NameResolver {
+	return &googleAnalyticsNameResolver{httpClient: httpClient, accountID: accountID}
+}
+
+func (r *googleAnalyticsNameResolver) ResolveInstanceName(ctx context.Context) (string, error) {
+	if r.accountID == "" {
+		return "", nil
+	}
+
+	endpoint, err := url.JoinPath("https://"+googleAnalyticsAPIHost, "v1alpha", "accounts", url.PathEscape(r.accountID))
+	if err != nil {
+		return "", fmt.Errorf("cannot build google analytics account URL: %w", err)
+	}
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
+	if err != nil {
+		return "", fmt.Errorf("cannot create google analytics account request: %w", err)
+	}
+
+	req.Header.Set("Accept", "application/json")
+
+	httpResp, err := r.httpClient.Do(req)
+	if err != nil {
+		return "", fmt.Errorf("cannot execute google analytics account request: %w", err)
+	}
+
+	defer func() {
+		_ = httpResp.Body.Close()
+	}()
+
+	// A non-2xx (revoked token, renamed/deleted account) is terminal: keep the
+	// generic source name rather than retry forever.
+	if httpResp.StatusCode < 200 || httpResp.StatusCode >= 300 {
+		return "", nil
+	}
+
+	var resp struct {
+		DisplayName string `json:"displayName"`
+	}
+	if err := json.NewDecoder(httpResp.Body).Decode(&resp); err != nil {
+		return "", fmt.Errorf("cannot decode google analytics account response: %w", err)
+	}
+
+	return resp.DisplayName, nil
+}
